Clean up index.js, created new EventListener file
